refactor(customer-analytics): extract accountsColumnConfigLogic
accountsLogic.ts grew past 700 lines once the column configurator landed.
Pull the column-config surface into its own logic so accountsLogic can
stay focused on filters, pagination, sorting, and role updates.

accountsColumnConfigLogic owns:
- `ACCOUNTS_HOGQL_PINNED_SELECT` / `ACCOUNTS_HOGQL_DEFAULT_SELECT`
  constants and the `buildAccountColumnGroups` resolver
- the column selection state (`selectColumns`, `columnConfiguratorVisible`),
  reorder / save / reset actions, and the
  `savedColumnConfiguration` loader
- the `visibleColumnNames` and `accountsColumnGroups` selectors
- afterMount: loads the saved configuration and hydrates the database
  schema if needed

accountsLogic connects to it via `connect()` so callers that already
read `selectColumns` or `visibleColumnNames` via accountsLogic keep
working, and the existing `clearSortIfColumnRemoved` listener still
fires when the user removes a sorted column.

Drive-by: the test file's `hogqlQuery.source.orderBy` assertions
TS-errored because `DataTableNode['source']` is a wide union; added a
local `orderByOf` helper that narrows to AccountsQuery so the
existing intent reads cleanly.

Generated-By: PostHog Code
Task-Id: 204010eb-282e-4463-9ba1-7aafe6e4b750
